adicionando filtros

+++ src/main/java/com/controledefinancapessoal/model/Receita.java
@@ -56,7 +56,7 @@ import java.util.Date;
             return valor;
         }
 
-        public void setValor(int valor) {
+        public void setValor(double valor) {
             this.valor = valor;
         }
